requirements.txt
aniso8601==8.0.0
astroid==2.4.2
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.11.0
pytz==2020.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==0.16.1
wrapt==1.12.1
